+++ deltafi-ui/src/App.vue
@@ -14,6 +14,9 @@
     </div>
     <FooterBanner />
     <Toast position="bottom-right" />
+    <Dialog v-model:visible="idle" header="Are you still there?" :modal="true" :style="{ width: '25vw' }">
+      <p>All background network activity has been paused due to user inactivity. Move the mouse to resume.</p>
+    </Dialog>
   </div>
 </template>
 
@@ -23,12 +26,13 @@ import AppTopBar from "@/components/header/AppTopBar";
 import AppMenu from "@/components/sidebar/AppMenu";
 import FooterBanner from "@/components/footer/FooterBanner";
 import Toast from "primevue/toast";
+import Dialog from "primevue/dialog";
 import { useRoute } from "vue-router";
 import useUiConfig from "@/composables/useUiConfig";
 import useServerSentEvents from "@/composables/useServerSentEvents";
 import useNotifications from "@/composables/useNotifications";
 import { computed, onBeforeMount, watch, nextTick, onMounted, onBeforeUnmount, provide, ref } from "vue";
-import { useTitle } from "@vueuse/core";
+import { useTitle, useIdle } from "@vueuse/core";
 import { usePrimeVue } from "primevue/config";
 
 const primevue = usePrimeVue();
@@ -37,9 +41,11 @@ const title = useTitle("DeltaFi");
 const notify = useNotifications();
 const { serverSentEvents } = useServerSentEvents();
 const { uiConfig, fetchUiConfig } = useUiConfig();
-
+const idleTimeOut = (15 * 60 * 1000); // 15 min
+const { idle } = useIdle(idleTimeOut);
 const sidebarHidden = ref(false);
-provide('sidebarHidden', sidebarHidden);
+provide("sidebarHidden", sidebarHidden);
+provide("isIdle", idle);
 
 const sidebarClasses = computed(() => {
   return sidebarHidden.value ? "col sidebar hidden" : "col sidebar";
@@ -47,11 +53,11 @@ const sidebarClasses = computed(() => {
 
 onBeforeMount(() => {
   fetchUiConfig();
-  provide('uiConfig', uiConfig);
+  provide("uiConfig", uiConfig);
 });
 
 onMounted(async () => {
-  primevue.config.locale.dateFormat = 'yy-mm-dd';
+  primevue.config.locale.dateFormat = "yy-mm-dd";
   await nextTick();
   window.addEventListener("resize", onResize);
   onResize();
@@ -65,7 +71,7 @@ const onResize = () => {
   var windowWidth = window.innerWidth;
 
   if (!sidebarHidden.value && windowWidth < 1024) {
-    sidebarHidden.value = true
+    sidebarHidden.value = true;
   }
 };
 
@@ -85,8 +91,7 @@ watch(
     setPageTitle();
   }
 );
-
-serverSentEvents.addEventListener('announcement', (event) => {
+serverSentEvents.addEventListener("announcement", (event) => {
   notify.info("System Announcement", event.data, 0);
 });
 </script>